requirements.txt
# Core dependencies
streamlit>=1.31.0

# Document processing
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Vector database and embeddings
chromadb>=0.4.22
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
sentence-transformers>=2.3.0

# LLM providers
google-generativeai>=0.3.0
anthropic>=0.18.0
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0
